app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["vtitoff <v.titoff2010@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.8"}
asyncpg = "^0.30.0"
alembic = "^1.14.1"
black = "^25.1.0"
isort = "^6.0.0"
pydantic-settings = "^2.7.1"
sqlalchemy = "^2.0.38"
uvicorn = "^0.34.0"
werkzeug = "^3.1.3"
fastapi-pagination = "^0.12.34"
orjson = "^3.10.15"
redis = "^5.2.1"
pyjwt = "^2.10.1"
typer = "^0.15.2"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
aiohttp = "^3.11.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
